+++ src/main/kotlin/com/pswidersk/gradle/python/MinicondaSetupTask.kt
@@ -7,6 +7,9 @@ import org.gradle.process.ExecResult
 import java.io.File
 import java.net.URL
 import javax.inject.Inject
+import java.net.URLConnection
+import java.util.*
+
 
 abstract class MinicondaSetupTask @Inject constructor(
     private val execOperations: ExecOperations,
@@ -28,7 +31,13 @@ abstract class MinicondaSetupTask @Inject constructor(
         minicondaDirFile.mkdirs()
         val minicondaInstaller =
             minicondaDirFile.resolve("$DEFAULT_MINICONDA_RELEASE-${minicondaVersion.get()}-$os-$arch.$exec")
-        downloadMiniconda(minicondaInstaller)
+        downloadMiniconda(
+            minicondaRepoUrl,
+            minicondaRepoUsername,
+            minicondaRepoPassword,
+            minicondaRepoHeaders,
+            minicondaInstaller
+        )
         allowInstallerExecution(minicondaInstaller)
         logger.lifecycle("Installing $DEFAULT_MINICONDA_RELEASE...")
         execOperations.exec {
@@ -57,13 +66,35 @@ abstract class MinicondaSetupTask @Inject constructor(
         }
     }
 
-    private fun downloadMiniconda(minicondaFile: File) {
-        logger.lifecycle("Downloading $DEFAULT_MINICONDA_RELEASE to: ${minicondaFile.canonicalPath}")
-        val minicondaInputStream = URL("https://repo.anaconda.com/miniconda/${minicondaFile.name}").openStream()
+    private fun downloadMiniconda(
+        minicondaRepoUrl: String,
+        minicondaRepoUsername: String,
+        minicondaRepoPassword: String,
+        minicondaRepoHeaders: Map<String, String>,
+        minicondaFile: File
+    ) {
+        logger.lifecycle("Downloading $DEFAULT_MINICONDA_RELEASE to: ${minicondaFile.canonicalPath} from: $minicondaRepoUrl")
+        val connection = URL("${minicondaRepoUrl}/${minicondaFile.name}").openConnection()
+        if (minicondaRepoUsername.isNotBlank()) {
+            addBasicAuth(minicondaRepoUsername, minicondaRepoPassword, connection)
+        }
+        minicondaRepoHeaders.forEach { connection.addRequestProperty(it.key, it.value) }
+        val minicondaInputStream = connection.getInputStream()
         minicondaInputStream.use { inputStream ->
             minicondaFile.outputStream().use { outputStream ->
                 inputStream.copyTo(outputStream)
             }
         }
+    }
+
+    private fun addBasicAuth(
+        minicondaRepoUsername: String,
+        minicondaRepoPassword: String,
+        connection: URLConnection
+    ) {
+        logger.lifecycle("Adding basic authorization headers for '$minicondaRepoUsername' user.")
+        val userAndPass = "$minicondaRepoUsername:$minicondaRepoPassword"
+        val basicAuth = "Basic ${String(Base64.getEncoder().encode(userAndPass.toByteArray()))}"
+        connection.setRequestProperty("Authorization", basicAuth)
     }
 }